Programmable Block Placer: Fix placement offset and property issues
- The placement offset was possible to be programmed as 0, which meant the
  positions of the placer itself and thus prevented placing any blocks.
  The placement position is no longer cached, and it's
  now (placementOffset [0..15]) + 1.
- The configuration Property values in the placer are now stored as ints,
  so that the full range 0..255 of a byte is usable, and still also empty
  properties can be distinguished.
- Empty properties (value = -1) are now longer applied to TEs,
  this was an obvious oversight in the method that applies them.

+++ src/main/java/fi/dy/masa/autoverse/inventory/wrapper/machines/ItemHandlerWrapperPlacerProgrammable.java
@@ -12,7 +12,7 @@ public class ItemHandlerWrapperPlacerProgrammable extends ItemHandlerWrapperSequ
     private final SequenceMatcher sequenceMarkerHighBit;
     private final SequenceMatcherVariable sequenceTrigger;
     private final SequenceMatcherVariable[] propertySequences;
-    private final byte[] propertyValues;
+    private final int[] propertyValues;
     private int subState;
 
     public ItemHandlerWrapperPlacerProgrammable(
@@ -36,8 +36,8 @@ public class ItemHandlerWrapperPlacerProgrammable extends ItemHandlerWrapperSequ
             this.getSequenceManager().add(this.propertySequences[i]);
         }
 
-        this.propertyValues = new byte[this.propertySequences.length];
-        Arrays.fill(this.propertyValues, (byte) -1);
+        this.propertyValues = new int[this.propertySequences.length];
+        Arrays.fill(this.propertyValues, -1);
     }
 
     @Override
@@ -85,7 +85,7 @@ public class ItemHandlerWrapperPlacerProgrammable extends ItemHandlerWrapperSequ
 
         for (int id = 0; id < this.propertySequences.length; id++)
         {
-            this.propertyValues[id] = (byte) this.propertySequences[id].parseValueFromSequence(highBitMarker);
+            this.propertyValues[id] = this.propertySequences[id].parseValueFromSequence(highBitMarker);
         }
     }
 
@@ -150,7 +150,7 @@ public class ItemHandlerWrapperPlacerProgrammable extends ItemHandlerWrapperSequ
         tag = super.writeToNBT(tag);
 
         tag.setByte("SubState", (byte) this.subState);
-        tag.setByteArray("PropertyValues", this.propertyValues);
+        tag.setIntArray("PropertyValues", this.propertyValues);
 
         return tag;
     }
@@ -161,7 +161,7 @@ public class ItemHandlerWrapperPlacerProgrammable extends ItemHandlerWrapperSequ
         super.readFromNBT(tag);
 
         this.subState = tag.getByte("SubState");
-        byte[] props = tag.getByteArray("PropertyValues");
+        int[] props = tag.getIntArray("PropertyValues");
 
         for (int i = 0; i < props.length && i < this.propertyValues.length; i++)
         {
